+++ modules/core/test/math/testColVector.cpp
@@ -90,9 +90,9 @@ int main()
   {
     vpColVector v(4);
     std::vector<double> bench1(4);
-    for(size_t i=0; i<v.size(); i++) {
-      v[i] = i;
-      bench1[i] = i;
+    for(unsigned int i=0; i<v.size(); i++) {
+      v[i] = (double)i;
+	  bench1[i] = (double)i;
     }
     if (test("v", v, bench1) == false)
       return err;
@@ -112,7 +112,7 @@ int main()
 
     vpColVector r1;
     for(size_t i=0; i<4; i++)
-      r1.stack(i);
+      r1.stack((double)i);
 
     vpColVector r2 = r1.extract(1, 3);
     if (test("r2", r2, bench3) == false)
